requirements.txt
feedparser==6.0.12
python-dotenv==1.2.1
stoat.py==1.2.1